Redid server again, firebase specific file is gone. Integrated into App.js
